Rename `storages` to `adapters` (#194)
