+++ src/adapters/_stringify.js
