+++ src/adapters/browser.js
