+++ src/adapters/file-async.js
