+++ src/adapters/file-sync.js
