+++ src/adapters/memory.js
